+++ src/app/components/select/demos/basic.ts
@@ -6,7 +6,7 @@ import { FormGroup, FormBuilder, Validators } from '@angular/forms';
   template: `
     <form [formGroup]="form" (ngSubmit)="submit()" ntFormAutofocus>
       <nt-form-field label="单选">
-        <nt-select name="single" placeholder="单选" formControlName="single">
+        <nt-select name="single" placeholder="单选" formControlName="single" filter>
           <nt-option *ngFor="let alphabet of alphabets" [value]="alphabet">
             {{alphabet}}
           </nt-option>